python/msgpack-reader-py/pyproject.toml
[tool.poetry]
name = "msgpack-reader-py"
version = "0.1.0"
description = ""
authors = ["Artavazd Balaian <reasyu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
msgpack-python = "^0.5.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
ruff = "^0.4.2"
msgpack-types = "^0.2.0"
pytest = "^8.2.0"
coverage = "^7.5.0"
pyperf = "^2.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]


[tool.mypy]
mypy_path = "./"
explicit_package_bases = true
ignore_missing_imports = true